+++ packages/client-common/src/index.ts
@@ -166,8 +166,10 @@ export {
 } from "./utils";
 export { LogWriter, DefaultLogger, type LogWriterParams } from "./logger";
 export { getCurrentStackTrace, enhanceStackTrace } from "./error";
+export { COMPRESSION_METHODS } from "./connection";
 export type {
   CompressionSettings,
+  CompressionMethod,
   Connection,
   ConnectionParams,
   ConnInsertResult,